+++ gospatial/app/router.go
@@ -6,8 +6,6 @@ import (
 	"net/http"
 )
 
-import mylogger "gospatial/logs"
-
 
 // Router for http api calls
 func Router() *mux.Router {
@@ -15,7 +13,7 @@ func Router() *mux.Router {
 	for _, route := range routes {
 		var handler http.Handler
 		// log.Println("Attaching HTTP handler for route:", route.Method, route.Pattern)
-		mylogger.Logger.Info("Attaching HTTP handler for route: ", route.Method, route.Pattern)
+		ServerLogger.Info("Attaching HTTP handler for route: ", route.Method, route.Pattern)
 		handler = route.HandlerFunc
 		router.
 			Methods(route.Method).