+++ gospatial/app/tilelayerHandlers.go
@@ -5,14 +5,12 @@ import (
 	"net/http"
 )
 
-import mylogger "gospatial/logs"
-
 // NewLayerHandler creates a new geojson layer. Saves layer to database and adds layer to customer
 // @param apikey
 // @return json
 func NewTileLayerHandler(w http.ResponseWriter, r *http.Request) {
 	// networkLoggerInfoIn.Printf("%v\n", r)
-	mylogger.Network.Debug(r)
+	NetworkLogger.Debug(r)
 
 	// Get params
 	apikey := r.FormValue("apikey")
@@ -30,7 +28,7 @@ func NewTileLayerHandler(w http.ResponseWriter, r *http.Request) {
 	// Check for apikey in request
 	if apikey == "" {
 		// networkLoggerError.Println(r.RemoteAddr, "POST /api/v1/tilelayer [401]")
-		mylogger.Network.Error(r.RemoteAddr,  " POST /api/v1/tilelayer [401]")
+		NetworkLogger.Error(r.RemoteAddr,  " POST /api/v1/tilelayer [401]")
 		http.Error(w, `{"status": "fail", "result": "unauthorized"}`, http.StatusUnauthorized)
 		return
 	}
@@ -39,7 +37,7 @@ func NewTileLayerHandler(w http.ResponseWriter, r *http.Request) {
 	customer, err := DB.GetCustomer(apikey)
 	if err != nil {
 		// networkLoggerWarning.Println(r.RemoteAddr, "POST /api/v1/tilelayer [404]")
-		mylogger.Network.Error(r.RemoteAddr,  " POST /api/v1/tilelayer [404]")
+		NetworkLogger.Error(r.RemoteAddr,  " POST /api/v1/tilelayer [404]")
 		http.Error(w, err.Error(), http.StatusNotFound)
 		return
 	}
@@ -54,7 +52,7 @@ func NewTileLayerHandler(w http.ResponseWriter, r *http.Request) {
 	js, err := json.Marshal(data)
 	if err != nil {
 		// networkLoggerError.Println(r.RemoteAddr, "POST /api/v1/tilelayer [500]")
-		mylogger.Network.Critical(r.RemoteAddr,  " POST /api/v1/tilelayer [500]")
+		NetworkLogger.Critical(r.RemoteAddr,  " POST /api/v1/tilelayer [500]")
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
@@ -64,9 +62,9 @@ func NewTileLayerHandler(w http.ResponseWriter, r *http.Request) {
 	// allow cross domain AJAX requests
 	w.Header().Set("Access-Control-Allow-Origin", "*")
 	// networkLoggerInfo.Println(r.RemoteAddr, "POST /api/v1/layer [200]")
-	mylogger.Network.Info(r.RemoteAddr,  " POST /api/v1/tilelayer [200]")
+	NetworkLogger.Info(r.RemoteAddr,  " POST /api/v1/tilelayer [200]")
 	// networkLoggerInfoOut.Println(string(js))
-	mylogger.Network.Debug(js)
+	NetworkLogger.Debug(js)
 	w.Write(js)
 
 }