+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## [1.10.4] - 2016-08-05
 ### Added
  - switch logging to seelog
+ - refactor or logging
 ### Fixed
  - json error messages jsend complient (https://labs.omniti.com/labs/jsend)